Update 1.jpg
Replaced Custom Text image with The original Text image exported with the correct settings. The reason it was custom in the first place was an attempt to keep the same image quality without making the file size larger. however I did not succeed in pacing the text correctly so we a am setting it back to the OG file with the disadvantage of larger file size as to keep the same quality it had to be exported at 100% quality. If the original uncompressed image is still available please replace this one with it exported at 90% quality, progressive turned off and subsampling Chroma Quarter for the OG file size and complete compatibility